feat(atoms): introduce `OtpErlangAtomLarge`
I’m trying to reproduce a binary term from JS to match something
a real Erlang program produces and I found that even though my
atoms in question are < 255 bytes, they are encoded as large atoms.

To make this work, I introduced a new class `OtpErlangAtomLarge`
that encodes as a large atom, regardless of the size of its
contained identifier.